DeepestScatter_Train/requirements.txt
absl-py==0.5.0
appdirs==1.4.3
astor==0.7.1
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
blessings==1.7
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cmake-converter==1.2.0
colorama==0.4.0
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.2.3
Flask==1.0.2
future==0.17.1
gast==0.2.0
glog==0.3.1
graphviz==0.10.1
grpcio==1.15.0
h5py==2.8.0
hypothesis==3.82.1
idna==2.7
ipydatawidgets==4.0.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipyvolume==0.5.1
ipywebrtc==0.4.1
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lmdb==0.94
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
mock==3.0.3
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.16.3
onnx==1.3.0
onnx-caffe2==1.0.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
pick==0.6.4
pickleshare==0.7.5
Pillow==5.3.0
pipdate==0.3.2
progressbar2==3.38.0
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.6.1
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-gflags==3.1.2
python-nvd3==0.15.0
python-slugify==1.2.6
python-utils==2.3.0
pythreejs==2.0.0
pytils==0.3
pytz==2018.9
PyWavelets==1.0.1
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
readchar==2.0.1
requests==2.20.0
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.12
tensorboardX==1.6
tensorflow==1.12
tensorflow-estimator==1.13
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tools==0.1.9
toolz==0.9.0
torch==1.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
traittypes==0.2.1
tuna==0.2.0
typing==3.6.6
typing-extensions==3.6.6
Unidecode==1.0.22
urllib3==1.24
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
windows-curses==1.0
